+++ src/scheduledDevice.hpp
@@ -7,5 +7,5 @@
 class scheduledDevice : public device {
 public:
     uint8_t          deviceID;
-    virtual uint32_t dispatchEvent(uint8_t index, uint8_t data[4]) = 0;
+    virtual uint64_t dispatchEvent(uint8_t index, uint8_t data[4]) = 0;
 };
